+++ src/auth/interfaces/signin-event-body.interface.ts
@@ -1,4 +1,5 @@
 import { AuthStrategy } from "../enums/auth-type.enum";
+import { AuthDTO } from "./signup-event-body.interface";
 
 export interface Tokens {
     accessToken: string;
@@ -8,17 +9,4 @@ export interface Tokens {
 export interface SigninEventBody {
     auth: AuthDTO;
     tokens: Tokens;
-}
-export interface AuthDTO {
-    id: string;
-    createdAt: string; // ISO date
-    updatedAt: string; // ISO date
-    deletedAt: string | null;
-    uid: string;
-    strategy: AuthStrategy; // extend if needed
-    isPrimary: boolean;
-    isVerified: boolean;
-    isActive: boolean;
-    meta: Record<string, any> | null;
-    lastUsedAt: string; // ISO date
 }